In #2279 there was some debate on whether it is better to have each thread in a Metal SIMD group process whole `Q4_0` or `Q4_1` blocks, or have them process half blocks. 

On my M2 Max with 30-core GPU having threads process half a block is definitely faster, see table below. I'm curious to see if this change also improves (or worsens) performance on M1 Pro/Max.

TG-128 in ms/token on 30-core M2 Max. The command used was
```
./main -m $model -p "I believe the meaning of life is" --ignore-eos -n 128 -s 1234 --no-mmap -t 8 -ngl 1
```
 
| Quantization |  t/s Master | t/s PR | speedup |
|--:|--:|--:|--:|
|Q4_0 - 7B | 18.6 | 17.6 | 5.7% |
|Q4_1 - 7B | 20.0 | 18.9 | 5.8% |
|Q4_0 - 13B | 32.5 | 30.1 | 8.0% |
|Q4_1 - 13B | 33.7 | 32.2 | 4.7% |
|Q4_0 - 33B | 76.5 | 67.2 | 11.4% |
|Q4_1 - 33B | 82.3 | 75.3 | 9.3% |
|Q4_0 - 65B | 146.8 | 133.4 | 10.0% |
|Q4_1 - 65B | 152.8 | 145.4 | 5.1%