@lshzh-ww 2023-07-24
Test on M1 Max 32c

Quantization | t/s Master | t/s PR | speedup
-- | -- | -- | --
Q4_0 - 7B | 19.18 | 18.07 | 6.1%
Q4_0 - 33B | 73.9 | 70.9 | 4.2%

The performance improvement is solid. I am ashamed that I didn't test intensively last time; we would have lost quite a bit of performance if it weren't for you.