@ikawrakow 2023-07-24
>  I am ashamed that I didn't test intensively last time; we would have lost quite a bit of performance

@lshzh-ww That doesn't make sense. Your contribution in #2188 and #2248 were absolutely essential for improving Metal performance. Before you came along with #2188 we were basically stuck at 21.5 ms/t on M2 Max for the 7B model. This kind of collaborative progress is the whole point of open source!