@ikawrakow 2023-07-25
> @ikawrakow and @lshzh-ww
> 
> I assume your tables show `ms/t` instead of `t/s`, correct?

Yes.